refactored compression code
